# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.BatchCreateTestCasesRequest:
    properties:
      problem_id:
        type: string
      test_cases:
        items:
          $ref: '#/definitions/models.TestCase'
        minItems: 1
        type: array
    required:
    - problem_id
    - test_cases
    type: object
...
  models.CreateTestCaseRequest:
    properties:
      input:
        type: string
      is_sample:
        type: boolean
      memory_limit:
        description: 可选的内存限制（MB）
        type: integer
      output:
        type: string
      problem_id:
        type: string
      score:
        description: 用例分数
        type: integer
      time_limit:
        description: 可选的超时配置（ms）
        type: integer
    required:
    - input
    - output
    - problem_id
    type: object
...
  models.Problem:
    properties:
      ac_count:
        description: AC 次数
        type: integer
      author:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      difficulty:
        $ref: '#/definitions/models.ProblemDifficulty'
      hint:
        type: string
      id:
        type: string
      input:
        type: string
      is_public:
        type: boolean
      memory_limit:
        description: 内存限制(MB)
        type: integer
      output:
        type: string
      sample_input:
        type: string
      sample_output:
        type: string
      source:
        type: string
      status:
        $ref: '#/definitions/models.ProblemStatus'
      submit_count:
        description: 提交次数
        type: integer
      tags:
        items:
          type: string
        type: array
      time_limit:
        description: 时间限制(ms)
        type: integer
      title:
        type: string
      updated_at:
        type: string
    required:
    - description
    - title
    type: object
  models.ProblemDifficulty:
    enum:
    - easy
    - medium
    - hard
    type: string
    x-enum-comments:
      DifficultyEasy: 简单
      DifficultyHard: 困难
      DifficultyMedium: 中等
    x-enum-descriptions:
    - 简单
    - 中等
    - 困难
    x-enum-varnames:
    - DifficultyEasy
    - DifficultyMedium
    - DifficultyHard
  models.ProblemStatus:
    enum:
    - draft
    - published
    - archived
    type: string
    x-enum-comments:
      StatusArchived: 已归档
      StatusDraft: 草稿
      StatusPublished: 已发布
    x-enum-descriptions:
    - 草稿
    - 已发布
    - 已归档
    x-enum-varnames:
    - StatusDraft
    - StatusPublished
    - StatusArchived
...
  models.TestCase:
    properties:
      created_at:
        type: string
      id:
        type: string
      input:
        type: string
      is_sample:
        type: boolean
      memory_limit:
        description: 内存限制(MB)，可选
        type: integer
      output:
        type: string
      problem_id:
        type: string
      score:
        description: 用例分数（可选，用于部分分）
        type: integer
      time_limit:
        description: 可选的超时配置（优先级高于题目默认配置）
        type: integer
    type: object
...
  models.UpdateTestCaseRequest:
    properties:
      input:
        type: string
      is_sample:
        type: boolean
      memory_limit:
        type: integer
      output:
        type: string
      score:
        type: integer
      time_limit:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@zkcodearena.com
    name: API Support
    url: http://localhost/support
  description: |-
    ZK Code Arena 在线编程平台 REST API 文档
    提供用户管理、题库管理、班级课程管理、代码评测等功能
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://localhost/terms/
  title: ZK Code Arena API
  version: "1.0"
paths:
...
  /problem/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定题目（仅管理员）
      parameters:
      - description: 题目ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的题目ID
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 删除失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除题目（管理员）
      tags:
      - 题目
    get:
      consumes:
      - application/json
      description: 根据题目ID获取完整题目信息
      parameters:
      - description: 题目ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 题目详情
          schema:
            $ref: '#/definitions/models.Problem'
        "400":
          description: 无效的题目ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 需要登录
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 题目不存在
          schema:
            additionalProperties: true
            type: object
      summary: 获取题目详情
      tags:
      - 题目
    put:
      consumes:
      - application/json
      description: 更新题目信息（管理员或题目创建者）
      parameters:
      - description: 题目ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新的题目信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Problem'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/models.Problem'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 需要登录
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 题目不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 更新失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新题目（教师/管理员）
      tags:
      - 题目
...
  /testcase:
    post:
      consumes:
      - application/json
      description: 为指定题目创建新的测试用例
      parameters:
      - description: 测试用例信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateTestCaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/models.TestCase'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 创建失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建测试用例
      tags:
      - 测试用例
  /testcase/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定测试用例
      parameters:
      - description: 测试用例ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的测试用例ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 删除失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除测试用例
      tags:
      - 测试用例
    get:
      consumes:
      - application/json
      description: 根据测试用例ID获取详细信息
      parameters:
      - description: 测试用例ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 测试用例详情
          schema:
            $ref: '#/definitions/models.TestCase'
        "400":
          description: 无效的测试用例ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 获取失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取测试用例详情
      tags:
      - 测试用例
    put:
      consumes:
      - application/json
      description: 更新指定测试用例的信息
      parameters:
      - description: 测试用例ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTestCaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/models.TestCase'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 测试用例不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 更新失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新测试用例
      tags:
      - 测试用例
  /testcase/batch:
    post:
      consumes:
      - application/json
      description: 批量导入题目测试用例（教师/管理员）
      parameters:
      - description: 批量测试用例
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BatchCreateTestCasesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 需要登录
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 题目不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 创建失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 批量创建测试用例
      tags:
      - 测试用例
  /testcase/problem/{problem_id}:
    get:
      consumes:
      - application/json
      description: 根据题目ID获取所有测试用例
      parameters:
      - description: 题目ID
        in: path
        name: problem_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 测试用例列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的题目ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 获取失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取题目测试用例列表
      tags:
      - 测试用例
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"